jobs/release_scraping/requirements.txt
click==8.0.4
pytest==7.4.4
pytest-black==0.3.11
pytest-flake8==1.0.6
requests==2.32.3
beautifulsoup4==4.14.2
py==1.11.0
flake8==4.0.1
google-cloud-storage==2.18.2
fsspec==2024.2.0
